Revert "Add cloud configuration"
