--- backend/src/main/java/com/zuzhi/backend/rating/RatingRepository.java
@@ -1,21 +1,0 @@
-package com.zuzhi.backend.rating;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-/**
- * RatingRepository
- *
- * @author zuzhi
- * @date 14/03/2018
- */
-interface RatingRepository extends JpaRepository<Rating, Long> {
-    /**
-     * Returns all ratings with the given book id
-     *
-     * @param bookId book id
-     * @return all ratings with the given book id
-     */
-    List<Rating> findRatingsByBookId(Long bookId);
-}
