+++ backend/src/main/java/com/zuzhi/springbootvuejsproject/SpringBootVuejsProjectApplication.java
@@ -1,0 +1,18 @@
+package com.zuzhi.springbootvuejsproject;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * SpringBootVuejsProjectApplication
+ *
+ * @author zuzhi
+ * @date 13/03/2018
+ */
+@SpringBootApplication
+public class SpringBootVuejsProjectApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringBootVuejsProjectApplication.class, args);
+    }
+}
