+++ backend/src/main/java/com/zuzhi/springbootvuejsproject/book/BookRepository.java
@@ -1,0 +1,12 @@
+package com.zuzhi.springbootvuejsproject.book;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ * BookRepository
+ *
+ * @author zuzhi
+ * @date 13/03/2018
+ */
+interface BookRepository extends JpaRepository<Book, Long> {
+}
