+++ backend/src/main/java/com/zuzhi/springbootvuejsproject/rating/RatingController.java
@@ -1,0 +1,53 @@
+package com.zuzhi.springbootvuejsproject.rating;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Map;
+
+/**
+ * RatingController
+ *
+ * @author zuzhi
+ * @date 14/03/2018
+ */
+@RestController
+@RequestMapping("/ratings")
+public class RatingController {
+
+    private final RatingService ratingService;
+
+    @Autowired
+    public RatingController(RatingService ratingService) {
+        this.ratingService = ratingService;
+    }
+
+    @GetMapping
+    public List<Rating> findRatingsByBookId(@RequestParam(required = false, defaultValue = "0") Long bookId) {
+        if (bookId.equals(0L)) {
+            return ratingService.findAllRatings();
+        }
+        return ratingService.findRatingsByBookId(bookId);
+    }
+
+    @PostMapping
+    public Rating createRating(@RequestBody Rating rating) {
+        return ratingService.createRating(rating);
+    }
+
+    @DeleteMapping("/{ratingId}")
+    public void deleteRating(@PathVariable Long ratingId) {
+        ratingService.deleteRating(ratingId);
+    }
+
+    @PutMapping("/{ratingId}")
+    public Rating updateRating(@RequestBody Rating rating, @PathVariable Long ratingId) {
+        return ratingService.updateRating(rating, ratingId);
+    }
+
+    @PatchMapping("/{ratingId}")
+    public Rating updateRating(@RequestBody Map<String, String> updates, @PathVariable Long ratingId) {
+        return ratingService.updateRating(updates, ratingId);
+    }
+}
