+++ backend/src/main/java/com/zuzhi/springbootvuejsproject/rating/RatingService.java
@@ -1,0 +1,78 @@
+package com.zuzhi.springbootvuejsproject.rating;
+
+import com.google.common.base.Preconditions;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Map;
+
+/**
+ * RatingService
+ *
+ * @author zuzhi
+ * @date 14/03/2018
+ */
+@Service
+@Transactional(readOnly = true, rollbackFor = Exception.class)
+public class RatingService {
+
+    private final RatingRepository ratingRepository;
+
+    @Autowired
+    public RatingService(RatingRepository ratingRepository) {
+        this.ratingRepository = ratingRepository;
+    }
+
+    public List<Rating> findAllRatings() {
+        return ratingRepository.findAll();
+    }
+
+    public List<Rating> findRatingsByBookId(Long bookId) {
+        return ratingRepository.findRatingsByBookId(bookId);
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
+    public Rating createRating(Rating rating) {
+        Rating newRating = new Rating();
+        newRating.setBookId(rating.getBookId());
+        newRating.setStars(rating.getStars());
+        return ratingRepository.save(newRating);
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
+    public void deleteRating(Long ratingId) {
+        ratingRepository.deleteById(ratingId);
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
+    public Rating updateRating(Rating rating, Long ratingId) {
+        Preconditions.checkNotNull(rating);
+        Preconditions.checkState(rating.getId().equals(ratingId));
+        Preconditions.checkNotNull(ratingRepository.findById(ratingId));
+        return ratingRepository.save(rating);
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
+    public Rating updateRating(Map<String, String> updates, Long ratingId) {
+        final Rating rating = findRatingById(ratingId);
+        updates.keySet()
+                .forEach(key -> {
+                    switch (key) {
+                        case "stars":
+                            rating.setStars(Integer.parseInt(updates.get(key)));
+                            break;
+                        default:
+                            break;
+                    }
+                });
+        return ratingRepository.save(rating);
+    }
+
+    private Rating findRatingById(Long ratingId) {
+        return ratingRepository.findById(ratingId)
+                .orElseThrow(() -> new RatingNotFoundException("Rating not found. ID: " + ratingId));
+    }
+}
