+++ backend/src/main/resources/application.yml
@@ -1,0 +1,7 @@
+server:
+  port: 3000
+spring:
+  jackson:
+    property-naming-strategy: SNAKE_CASE
+    date-format: yyyy-MM-dd HH:mm:ss
+    time-zone: GMT+8
