+++ backend/src/test/java/com/zuzhi/springbootvuejsproject/SpringBootVuejsProjectApplicationTests.java
@@ -1,0 +1,16 @@
+package com.zuzhi.springbootvuejsproject;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class SpringBootVuejsProjectApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
