--- config/src/main/java/com/zuzhi/config/ConfigApplication.java
@@ -1,22 +1,0 @@
-package com.zuzhi.config;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.config.server.EnableConfigServer;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-/**
- * ConfigApplication
- *
- * @author zuzhi
- * @date 16/03/2018
- */
-@SpringBootApplication
-@EnableConfigServer
-@EnableEurekaClient
-public class ConfigApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ConfigApplication.class, args);
-    }
-}
