--- config/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-server.port=8081
-spring.application.name=config
-spring.cloud.config.server.git.uri=file:///${user.home}/application-config
-# eureka
-eureka.client.region=default
-eureka.client.registry-fetch-interval-seconds=5
-eureka.client.service-url.default-zone=http://discUser:discPass@localhost:8082/eureka/
-# security
-spring.security.user.name=configUser
-spring.security.user.password=configPass
-spring.security.user.roles=SYSTEM
