--- discovery/src/main/java/com/zuzhi/discovery/DiscoveryApplication.java
@@ -1,20 +1,0 @@
-package com.zuzhi.discovery;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-/**
- * DiscoveryApplication
- *
- * @author zuzhi
- * @date 16/03/2018
- */
-@SpringBootApplication
-@EnableEurekaServer
-public class DiscoveryApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DiscoveryApplication.class, args);
-	}
-}
