--- discovery/src/main/resources/bootstrap.properties
@@ -1,5 +1,0 @@
-# cloud config
-spring.cloud.config.name=discovery
-spring.cloud.config.uri=http://localhost:8081
-spring.cloud.config.username=configUser
-spring.cloud.config.password=configPass
