--- gateway/src/main/java/com/zuzhi/gateway/GatewayApplication.java
@@ -1,22 +1,0 @@
-package com.zuzhi.gateway;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-
-/**
- * GatewayApplication
- *
- * @author zuzhi
- * @date 16/03/2018
- */
-@SpringBootApplication
-@EnableZuulProxy
-@EnableEurekaClient
-public class GatewayApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GatewayApplication.class, args);
-    }
-}
