--- gateway/src/main/resources/bootstrap.properties
@@ -1,8 +1,0 @@
-# cloud config
-spring.cloud.config.name=gateway
-spring.cloud.config.discovery.service-id=config
-spring.cloud.config.discovery.enabled=true
-spring.cloud.config.username=configUser
-spring.cloud.config.password=configPass
-# eureka
-eureka.client.service-url.default-zone=http://discUser:discPass@localhost:8082/eureka/
